parser.py: Allow equal signs in field values
Previously, the parser did not account for the possibility of equal
signs (`=`) in the values of fields and would result in crashing the
script with a non descriptive error message, due to expecting a simple
`string.split("=")` to work.

However, it is completely possible to have equal signs in the value,
for example in an URL field: `URL = "https://example.org/query?x=1"`.

This adapts the parsing logic, making it stable against equal signs in
field values. Furthermore, it cleans up the `BibTeXEntry.from_string`
method a little bit and adds additional test cases to the unittests.

Fixes #3