Refactor parser to use string `entry_type`s
Previously, we used the `EntryType` enum to define the possible entry
types. This was a little too constrained, as it would fail parsing the
`BibTeXEntry`, if the `EntryType` was unknown. So in order to allow for
better usability of the tool, such as writing rules for entry types that
may be completely uncommon for my personal field, we refactor to use a
string `BibTeXEntry.entry_type` from now on.

This change needed to be carried out everwhere where `EntryType` was
used before, so it turned out to be a significant refactor of the code
base.

Fixes #5