Add new ruleset for `IEEEtran` citation style (#15)
Previously, we did not draw a precise distinction between the two
citation styles `IEEEtran` and `ieeetr` (from `IEEEtran.cls`). This
was mainly due to the reason that I did not know myself, that those
two were separate styles.

This adds a new ruleset for the LaTeX built-in `IEEEtran` citation
style. For this, we modify the way the script argument `ruleset` is
parsed by allowing to specify the rulesets shipped with the program
directly (e.g. by calling it by its name:
`bibtex_linter path/to/refs.bib IEEEtran`). However, the default
behavior stayed the same, so this change should be backward compatible.
Additionally, we adapt the `README.md` accordingly, drawing a more
precise disctinction between the `ieeetr` and `IEEEtran` styles.

Furthermore, this adds the observations for the styles: `plain`,
`apalike` and `IEEEtran` to the `test/test_template` directory.

Fixes #12