Fix inverted logic in `check_disallowed_field` (#19)
Previously, the logic of the `verification.check_disallowed_field`
method was inverted. Instead of returning a violation when the
disallowed field was present, it was returning the violation when it
was not present. My suspicion is that this was due to a copy/paste
error.

This fixes this bug, by returning a violation only when a disallowed
field is present.
The bug was first found in the logic of the linter rule of:
`ieeetran_rules.check_in_collection`.

Fixes #18